TONS of progress
I made a lot of progress, namely in the renderer, and the buffer system
that works a lot like the shader system.  I also have been adapting more
DXUT code to work with openGL, glfw, and my own systems.  the biggest
things to do next is to create a dialog/text manager for openGL, to
create the physics system, and to create the audio system.