Moved assets, and am beginning the fixing process
I am beginning the debugging process, and am thinking about maybe
changing the current config system to support more advanced XML
features, but for now I will be using it simply.  I added the rest of
the assets, but all art/models are going to be in a different zip.