Fixed font, and image loaders are done
I fixed an issue with the font, because i was trying to make a font
before I initialized freetype.  I fixed the JPG/PNG loaders, BUT i have
not tested to see if they work, as of now, they just don't throw any
errors.  It runs and shows a main menu, but it is messed up, and the
font is way too big.  I have suspicians that some things are not working
because of the GLI issue.