Fixed Transform Component,
I fixed the transform component to not try to use an uninitialized
matrix, and a couple other various things.  i removed libogg, and vorbis
because I am using SFML to load sounds.  I do have something render, but
it just looks like a red line right now, but HORRAY that is a step in
the right direction.  Now the assertion system calls assert() so i can
debug assertion failures as well.