New Material System and SPHERES
The sphere works, well one does, it seems not to like having multiple
spheres, but I am not sure what causes that.  If there are more than
one, it seems to combine their colors in some way.  The new material
system works in two ways, 1: a dummy created from memory, or 2: an xml
file that is loaded, but I am not sure how I will incorporate that into
the mesh loading process.